refactor(Moving functionality to upload images to its own hook.):
First step in the process of refactoring how images are uploaded
